refactor : 리팩토링 후 Deprecated 파일 정리 및 사용하지 않는 메소드 삭제
